requirements.txt
discord.py
asyncio
mechanicalsoup
dnspython
pymongo
imgurpython
async-cse
git+https://github.com/Rapptz/discord-ext-menus
jishaku